--- a/sharding/collation.go
+++ b/sharding/collation.go
@@ -113,11 +113,9 @@ func (c *Collation) CalculateChunkRoot() {
 
 // ConvertBackToTx converts raw blobs back to their original transactions.
 func ConvertBackToTx(rawBlobs []utils.RawBlob) ([]*types.Transaction, error) {
-
 	blobs := make([]*types.Transaction, len(rawBlobs))
 
 	for i := 0; i < len(rawBlobs); i++ {
-
 		blobs[i] = types.NewTransaction(0, common.HexToAddress("0x"), nil, 0, nil, nil)
 
 		err := utils.ConvertFromRawBlob(&rawBlobs[i], blobs[i])
@@ -129,10 +127,7 @@ func ConvertBackToTx(rawBlobs []utils.RawBlob) ([]*types.Transaction, error) {
 
 }
 
-// SerializeTxToBlob method serializes the input tx
-// and returns the blobs in byte array.
-func SerializeTxToBlob(txs []*types.Transaction) ([]byte, error) {
-
+func SerializeTxToRawBlob(txs []*types.Transaction) ([]*utils.RawBlob, error) {
 	blobs := make([]*utils.RawBlob, len(txs))
 	for i := 0; i < len(txs); i++ {
 		err := error(nil)
@@ -141,13 +136,24 @@ func SerializeTxToBlob(txs []*types.Transaction) ([]byte, error) {
 			return nil, fmt.Errorf("%v", err)
 		}
 	}
+	return blobs, nil
+}
+
+// SerializeTxToBlob method serializes the input tx
+// and returns the blobs in byte array.
+func SerializeTxToBlob(txs []*types.Transaction) ([]byte, error) {
+	blobs, err := SerializeTxToRawBlob(txs)
+	if err != nil {
+		return nil, fmt.Errorf("%v", err)
+	}
+
 	serializedTx, err := utils.Serialize(blobs)
 	if err != nil {
 		return nil, fmt.Errorf("%v", err)
 	}
 
 	if int64(len(serializedTx)) > collationSizelimit {
-		return nil, fmt.Errorf("The serialized body exceeded the collation size limit: %v", serializedTx)
+		return nil, fmt.Errorf("The serialized body size %v exceeded the collation size limit %v", len(serializedTx), collationSizelimit)
 	}
 
 	return serializedTx, nil
@@ -156,7 +162,6 @@ func SerializeTxToBlob(txs []*types.Transaction) ([]byte, error) {
 // DeserializeBlobToTx takes byte array blob and converts it back
 // to original txs and returns the txs in tx array.
 func DeserializeBlobToTx(serialisedBlob []byte) (*[]*types.Transaction, error) {
-
 	deserializedBlobs, err := utils.Deserialize(serialisedBlob)
 	if err != nil {
 		return nil, fmt.Errorf("%v", err)