--- a/sharding/utils/marshal.go
+++ b/sharding/utils/marshal.go
@@ -122,7 +122,7 @@ func Serialize(rawBlobs []*RawBlob) ([]byte, error) {
 
 // isSkipEvm is true if the first bit is 1
 func isSkipEvm(indicator byte) bool {
-	return indicator&skipEvmBits >> 7 == 1
+	return indicator&skipEvmBits>>7 == 1
 }
 
 // getDatabyteLength is calculated by looking at the last 5 bits.