### Results

Here's the numbers on my computer for the existing benchmark with 150 byte txs
```
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/sharding
BenchmarkSerialization10-12                30000             69008 ns/op           27805 B/op        353 allocs/op
BenchmarkSerialization100-12                5000            739374 ns/op          324040 B/op       3425 allocs/op
BenchmarkSerialization1000-12                300           5558360 ns/op         3127289 B/op      34044 allocs/op
BenchmarkSerialization10000-12                30          48292393 ns/op        63835295 B/op     160091 allocs/op
PASS
```

Here's the numbers for the equivalent benchmark on this branch:

```
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/sharding
BenchmarkSerializeRoundtrip10-12          100000             37949 ns/op           15103 B/op        210 allocs/op
BenchmarkSerializeRoundtrip100-12           3000            426936 ns/op          150057 B/op       2013 allocs/op
BenchmarkSerializeRoundtrip1000-12           500           3309168 ns/op         1483492 B/op      20020 allocs/op
PASS
```

_Note: I took out benchmark tests with 10k transactions because they were hitting the [collation size limit](https://github.com/prysmaticlabs/geth-sharding/blob/master/sharding/collation.go#L150) and failing silently._

The benchmark tests are now broken up into Serialization/Deserialization, and for each of those I included a benchmark of the process without RLP.

```
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/sharding
BenchmarkSerializeNoRLP10-12             1000000              3409 ns/op            2088 B/op         11 allocs/op
BenchmarkSerializeNoRLP100-12             100000             34454 ns/op           20880 B/op        101 allocs/op
BenchmarkSerializeNoRLP1000-12              5000            330732 ns/op          200609 B/op       1001 allocs/op
BenchmarkSerialize10-12                   100000             11328 ns/op            4890 B/op         52 allocs/op
BenchmarkSerialize100-12                   10000            131254 ns/op           48999 B/op        502 allocs/op
BenchmarkSerialize1000-12                   1000           1251594 ns/op          481066 B/op       5003 allocs/op
BenchmarkDeserialize10-12                  50000             25462 ns/op           10208 B/op        158 allocs/op
BenchmarkDeserialize100-12                 10000            246955 ns/op          101008 B/op       1511 allocs/op
BenchmarkDeserialize1000-12                  500           3257192 ns/op         1001747 B/op      15014 allocs/op
BenchmarkDeserializeNoRLP10-12            500000              4435 ns/op            2736 B/op         16 allocs/op
BenchmarkDeserializeNoRLP100-12           100000             33656 ns/op           26480 B/op        109 allocs/op
BenchmarkDeserializeNoRLP1000-12            5000            333305 ns/op          257520 B/op       1012 allocs/op
BenchmarkSerializeRoundtrip10-12           50000             44685 ns/op           15103 B/op        210 allocs/op
BenchmarkSerializeRoundtrip100-12           5000            409291 ns/op          150061 B/op       2013 allocs/op
BenchmarkSerializeRoundtrip1000-12           300           5520262 ns/op         1483490 B/op      20019 allocs/op
PASS
```

RLP encoding now consumes 70-80% of wall clock time when running these benchmarks. Optimizing RLP would be the next logical step, but according to the [wiki](https://github.com/ethereum/wiki/wiki/RLP), the geth team are already aware that its inefficient and in the process of addressing perf issues.

### Optimizations
In the existing code, blob serialization/deserialization appends encoded data to an array in a loop. When data's appended to an array at full capacity, go allocates a new array with enough capacity and copies all of the data over. At 10k transactions, this becomes pretty inefficient.

```
var data []byte
for i := 0; i < len(blobs); i++ {
  // this allocates a new array on every iteration!
  data = append(data, byte(i))
}
```

Instead, I created an array with enough capacity
```
data := make([]byte, 0, len(blobs)
for i:= 0; i < len(blobs); i++ {
  data = append(data, byte(i)
}
```

I'm sure there are other optimizations that could be made, but this seemed to be the most glaring one to me. The next optimization would be to avoid encoding/decoding with both RLP and blob serialization.

Edit: Updated with `-benchmem` numbers