@nisdas 2018-06-01
Hey thanks for this ! Really like how you have changed optimisations you have done in both the `Serialize` and `Deserialize` methods.  Everything is more readable and modular.  Great Job !