@terenc3t 2018-06-01
There have been conversations in [ethereum/sharding](https://gitter.im/ethereum/sharding?utm_source=share-link&utm_medium=link&utm_campaign=share-link) Gitter channel to use protobuf. I interested to see what performance gain we get instead of rlp. 70-80% is quite a lot.