@rawfalafel 2018-06-01
the build seems to be failing on travis but otherwise, yes! i'll take a look at it tomorrow morning (I'm in Japan time).