@terenc3t 2018-06-02
@rawfalafel Lint failed. Don't forget to run gofmt : )