@rawfalafel 2018-06-03
Thanks for the quick turnaround @prestonvanloon !

Btw, I noticed that PRs are getting merged directly into master. Have you discussed squashing the branch into a single commit instead?