@prestonvanloon 2018-06-03
@rawfalafel I think we did discuss this. It seems to be the default for our repo now. 

I think this is ready to squash and merge. @nisdas / @terenc3t do you have any other comments?

Thanks so much, Yutaro!